+++ src/components/pages/apps/genes/gsea/gsea-plot/gsea-graph-axes-props-panel.tsx
@@ -1,27 +1,29 @@
+import { IDisplayPlot } from '@/components/plot/axes/plot/axes-groups-popover'
 import { AxesDisplayPropsPanel } from '../../../../../plot/axes/plot/axes-props-panel'
 import { useGseaPlot } from './gsea-plot-provider'
 
 export function GseaGraphAxesPropsPanel() {
   const { pathways } = useGseaPlot()
 
-  return (
-    <AxesDisplayPropsPanel
-      plotIds={pathways.map((p) => ({ id: p.id, title: p.name }))}
-      axesGroups={[
-        {
-          id: 'es',
-          title: 'ES',
-          axesIds: [
-            { id: 'x', axis: 'x', title: 'X' },
-            { id: 'y', axis: 'y', title: 'Y' },
-          ],
-        },
-        {
-          id: 'snr',
-          title: 'SNR',
-          axesIds: [{ id: 'y', axis: 'y', title: 'Y' }],
-        },
-      ]}
-    />
-  )
+  const plots: IDisplayPlot[] = pathways.map((p) => ({
+    id: p.id,
+    title: p.name,
+    groups: [
+      {
+        id: 'es',
+        title: 'ES',
+        axes: [
+          { id: 'x', axis: 'x', title: 'X' },
+          { id: 'y', axis: 'y', title: 'Y' },
+        ],
+      },
+      {
+        id: 'snr',
+        title: 'SNR',
+        axes: [{ id: 'y', axis: 'y', title: 'Y' }],
+      },
+    ],
+  }))
+
+  return <AxesDisplayPropsPanel plots={plots} />
 }