+++ src/components/pages/apps/genes/gsea/gsea-plot/gsea-settings-store.ts
@@ -82,7 +82,7 @@ export interface IGseaDisplayProps {
   }
   //title: ITextProps & { offset: number }
   page: {
-    scale: number
+    //scale: number
     columns: number
   }
   plot: {
@@ -114,7 +114,7 @@ export interface IGseaDisplayProps {
 export const DEFAULT_GSEA_DISPLAY_PROPS: IGseaDisplayProps = {
   page: {
     columns: 3,
-    scale: 1,
+    //scale: 1,
   },
   phenotypes: {
     invert: false,